Реализован подсчет времени выполнения программы

+++ lw1/Zaykov Maxim/MonteCarlo/MonteCarlo/RandomGenerator.h
@@ -2,6 +2,6 @@
 class RandomGenerator
 {
 public:
-	double getRandomDouble(const double & min, const double & max);
+	double getRandomDouble(const int & min, const int & max);
 };
